format args
